refactor: layout text (#372)
## Description
It appears that there are two areas where mistakes are causing this bug,
one is the selection of break opportunities not being in accordance with
the rules(https://www.unicode.org/reports/tr14), and the other is
setting the wrong minWidth. I have removed the logic related to minWidth
and currently, it doesn't seem to have any issues, I would like to know
if there are any risks that I have not noticed here. 🤔️

After fix:


https://user-images.githubusercontent.com/22126563/212294794-0c48984f-12b4-4436-9013-f0ef256cb150.mov

Before fix:


![image](https://user-images.githubusercontent.com/22126563/215529627-e6c1b261-f93f-40dc-99b1-5269391dc906.png)


## Update tests
I made a couple of adjustments to the test snapshots.
#### 1

![image](https://user-images.githubusercontent.com/22126563/215499965-d786036d-e445-42df-b5e5-ecc7a8c84382.png)

![image](https://user-images.githubusercontent.com/22126563/215499833-eb484d2f-7aaf-4cfa-96ab-9fa7b1017bbc.png)

## Notice:
Currently, the selection of line-break points is significantly more
consistent with the original HTML representation compared to before the
fix, but it is not 100% identical. There is still a possibility of some
differences remaining.

#### Closes: #360, Closes: #386